requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
cffi==1.15.1
channels==3.0.4
channels-redis==4.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.2
daphne==3.0.2
Django==4.0
django-bootstrap4==23.2
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.1
msgpack==1.0.5
Pillow==10.0.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
service-identity==23.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
